Setup alert messages and logs
